Fixed map pin structure Id
javalos noticed that the belong chicago event map had stopped working.  No js or vtl errors, just no events were being put into the script.  Turns out the map pin structure id was wrong.  I must have changed the id to the wrong one and not changed the name of the macro, so it continued to work with the cached version.  The cache must have recently cleared, making the check for the structure fail.  As an aside about liferay and this ID madness:  When you list out the structures at admin->content; manage->structures, there is an ID column.  On prod, the map pin structure ID is 565902.  If you click on the structure link, there is a field called 'Structure Key.' On prod, the map pin structure key is 565901.  The function getStructureId returns 565901...

Next step is to look into the uuid field, although the functions I've seen so far that take a uuid also take a groupId... making me wonder what a uuid actually is in Liferay...